+++ src/model1/StayedStudent.java
@@ -1,16 +1,19 @@
 package model1;
 
+import sinhVienDangKy.TakeData;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class StayedStudent {
-    private Map<String, Student> studentMap;
+    private static List<Student> students = new ArrayList<>();
+    public TakeData layDuLieuSV;
+    private Map<String, Student> studentMap = new HashMap<>();
 
     public StayedStudent() {
-        this.studentMap = new HashMap<>();
-
+        this.students = new ArrayList<>();
+        this.layDuLieuSV = TakeData.getInstances();
         Student st1 = new Student("Nguyễn Văn A", "23130001", "Nam", "Công nghệ thông tin", "24/01/2005", "A", "A04", "Bình Định", "123456", "0987654321", "Kinh", "Con liệt sĩ, thương binh, bệnh binh");
         Student st2 = new Student("Nguyễn Thị B", "23130002", "Nữ", "Công nghệ sinh học", "20/05/2004", "D", "D10", "Tiền Giang", "234567", "0345678990", "Mông", "Gia đình đặc biệt khó khăn");
         Student st3 = new Student("Nguyễn Văn C", "23130003", "Nam", "Công nghệ thực phẩm", "02/10/2005", "C", "C02", "Long An", "341678", "0168390591", "Kinh", "");
@@ -19,29 +22,54 @@ public class StayedStudent {
         Student st6 = new Student("Đinh Thị M", "23130014", "Nữ", "Thú y", "19/05/2006", "E", "E04", "Kiên Giang", "428450", "0470646432", "Thái", "");
         Student st7 = new Student("Trần Văn N", "23130015", "Nam", "Kinh tế", "17/01/2003", "F", "F06", "TP.HCM", "085342", "0470646289", "Kinh", "Gia đình đặc biệt khó khăn");
         Student st8 = new Student("Đặng Thị O", "23130021", "Nữ", "Ngôn ngữ anh", "09/11/2001", "B", "B11", "Long An", "581534", "0260641237", "Kinh", "");
+        List<Student> storedData = layDuLieuSV.getsVLuu();
 
-        addStudent(st1);
-        addStudent(st2);
-        addStudent(st3);
-        addStudent(st4);
-        addStudent(st5);
-        addStudent(st6);
-        addStudent(st7);
-        addStudent(st8);
-
-    }
-
-    // Thêm sinh viên vào danh sách
-    public void addStudent(Student student) {
-        if (student == null || student.getMssv() == null) {
-            throw new IllegalArgumentException("Thông tin sinh viên không hợp lệ!");
+        students.add(st1);
+        students.add(st2);
+        students.add(st3);
+        students.add(st4);
+        students.add(st5);
+        students.add(st6);
+        students.add(st7);
+        students.add(st8);
+        if (storedData != null && !storedData.isEmpty()) {
+            students.addAll(storedData);
+        } else {
+            System.out.println(" ");
         }
-        studentMap.put(student.getMssv(), student);
+
+        for (Student student : students) {
+            studentMap.put(student.getMssv(), student);
+        }
+
     }
 
-    // Tìm sinh viên theo MSSV
-    public Student findStudentById(String mssv) {
-        return studentMap.get(mssv.trim());
+
+    public Student getStudentById(String id) {
+        for (Student s : students) {
+            if (s.getMssv().equals(id)) {
+                return s;
+            }
+        }
+        return null;
+    }
+
+    public static List<Student> getStudents() {
+        return students;
+    }
+
+    public void setStudents(List<Student> students) {
+        this.students = students;
+    }
+
+    // Tim sinh vien theo mssv
+    public List<Student> searchStudentByMSSV(String mssv) {
+        List<Student> result = new ArrayList<>();
+        Student foundStu = studentMap.get(mssv.trim());
+        if (foundStu != null) {
+            result.add(foundStu);
+        }
+        return result;
     }
 
     // Cap nhat thong tin sinh vien
@@ -49,21 +77,37 @@ public class StayedStudent {
         if (updatedStudent == null || updatedStudent.getMssv() == null) {
             throw new IllegalArgumentException("Thông tin sinh viên không hợp lệ!");
         }
-
-        String mssv = updatedStudent.getMssv();
-        if (studentMap.containsKey(mssv)) {
-            studentMap.put(mssv, updatedStudent); // Thay thế thông tin sinh viên trong map
-        } else {
-            throw new IllegalArgumentException("Không tìm thấy sinh viên có MSSV: " + mssv);
+        boolean isUpdated = false;
+        for (int i = 0; i < students.size(); i++) {
+            if (students.get(i).getMssv().equals(updatedStudent.getMssv())) {
+                students.set(i, updatedStudent); // Cập nhật thông tin
+                isUpdated = true;
+                break;
+            }
+        }
+        if (!isUpdated) {
+            throw new IllegalArgumentException("Không tìm thấy sinh viên có MSSV: " + updatedStudent.getMssv());
         }
     }
 
-    // Xóa sinh viên khỏi danh sách theo MSSV
-    public boolean removeStudentById(String mssv) {
-        return studentMap.remove(mssv.trim()) != null;
+    // Xoa sinh vien
+    public boolean removeStudentById(String studentID) {
+        return students.removeIf(student -> student.getMssv().equals(studentID));
     }
 
-    public Map<String, Student> getStudentMap() {
-        return studentMap;
+
+    public List<String> getStudentStrings() {
+        List<String> studentStrings = new ArrayList<>();
+        for (Student student : students) {
+            studentStrings.add(student.toString());
+        }
+        return studentStrings;
+    }
+
+    public Student getStudentAtRow(int rowIndex) {
+        if (rowIndex >= 0 && rowIndex < students.size()) {
+            return students.get(rowIndex);
+        }
+        return null;
     }
 }
