+++ src/sinhVienDangO/StudentListButtonListener.java
@@ -1,6 +1,6 @@
 package sinhVienDangO;
 
-import model.StudentController;
+import model1.StayedStudent;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
@@ -8,11 +8,11 @@ import java.awt.event.ActionListener;
 
 public class StudentListButtonListener implements ActionListener {
     private StudentView studentView;
-    private StudentController studentController;
+    private StayedStudent stayedStudent;
 
-    public StudentListButtonListener(StudentView studentView, StudentController studentController) {
+    public StudentListButtonListener(StudentView studentView, StayedStudent stayedStudent) {
         this.studentView = studentView;
-        this.studentController = studentController;
+        this.stayedStudent = stayedStudent;
     }
 
     @Override
@@ -20,11 +20,11 @@ public class StudentListButtonListener implements ActionListener {
         StudentListView studentListView = new StudentListView();
 
         DefaultListModel<String> listModel = new DefaultListModel<>();
-        for (String student : studentController.getStudentStrings()) { // Giả sử getStudents() trả về danh sách sinh viên
+        for (String student : stayedStudent.getStudentStrings()) { // Giả sử getStudents() trả về danh sách sinh viên
             listModel.addElement(student);
         }
 
-        studentListView.updateStudentList(studentController.getStudents());
+        studentListView.updateStudentList(stayedStudent.getStudents());
 
         studentListView.setVisible(true);
         studentView.setVisible(false);
